Point at latest WebIDE demo.
The demo was renamed and includes a number of improvements, including the ability to view colorized factory methods, refresh buttons, better menu placement and links to testing from the root namespace. It also has a number of regressions, such as missing home buttons on object inspectors, and lacking help and about links at the home page.