Fix  root namespace.
Similar fix was done for WebIDE, but not for Ampleforth. This compensates for the problem that the Hopscotch UI is no longer in the basic platform.  The wrapper platform that includes the UI is created by the app, but the namespace population code misbehaves in that case. This works around that.