Fix expat performance bug
The exponential slowdown we see over time when running the IDE seems to be due to the cost of inserting expats into the expat table, which slows down in the same way, because the maps in CollectionsForPrimordialSoup don't deal well with large numbers of integer keys.  This addresses that to a considerable degree, by replacing the Map with a List. It may not be perfect, time will tell, but it is certainly a big improvement.